examples/datazone-msk-governance/resources/lambda/requirements.txt
boto3
openlineage-python
pyyaml
numpy
aws-glue-schema-registry
kafka-python
aws-msk-iam-sasl-signer-python